requirements.txt
django==4.2.27
djangorestframework==3.16.1
python-dotenv==1.2.1
requests==2.32.5
python-dateutil==2.9.0.post0
black==25.11.0
ruff==0.14.9
